requirements.txt
annotated-types==0.7.0
dnspython==2.6.1
et-xmlfile==1.1.0
motor==3.5.1
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qrcode==7.4.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
